removed write and read service
