+++ src/Core/Shared/Services/AccessConditionService.cs
@@ -12,11 +12,6 @@ namespace Biobanks.Directory.Services
     {
         public AccessConditionService(BiobanksDbContext db) : base(db) { }
 
-
-    public override async Task<int> GetUsageCount(int id)
-        => await _db.Collections.CountAsync(x => x.AccessConditionId == id);
-
-
     protected override IQueryable<AccessCondition> Query()
             => Query().Include(x => x.Organisations);
     }
