+++ src/Core/Shared/Services/ReferenceDataService.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace Biobanks.Shared.Services
 {
-    public abstract class ReferenceDataService<T> : IReferenceDataService<T> where T : BaseReferenceData
+    public  class ReferenceDataService<T> : IReferenceDataService<T> where T : BaseReferenceData
     {
         protected readonly BiobanksDbContext _db;
 
@@ -27,7 +27,6 @@ namespace Biobanks.Shared.Services
                 .ToListAsync(); 
         public async Task<ICollection<T>> List()
             => await List(string.Empty);
-        public abstract Task<int> GetUsageCount(int id);
 
   }
 }
