+++ src/Submissions/Api/Services/Directory/Contracts/IMaterialTypeService.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Biobanks.Entities.Shared.ReferenceData;
@@ -11,6 +11,8 @@ namespace Biobanks.Submissions.Api.Services.Directory.Contracts
         Task<int> GetExtractionProcedureMaterialDetailsCount(string id);
 
         Task<IEnumerable<OntologyTerm>> GetMaterialTypeExtractionProcedures(int id, bool onlyDisplayable = false);
-    }
+        IEnumerable<string> ExtractDistinctMaterialTypes(Collection collection);
+
+  }
 }
