+++ src/Submissions/Api/Services/Directory/Contracts/ISampleSetService.cs
@@ -1,3 +1,4 @@
+using Biobanks.Entities.Data;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -5,10 +6,13 @@ namespace Biobanks.Submissions.Api.Services.Directory.Contracts;
 
 public interface ISampleSetService
 {
+  Task AddSampleSetAsync(SampleSet sampleSet);
+  Task UpdateSampleSetAsync(SampleSet sampleSet);
+  Task DeleteSampleSetAsync(int id);
 
   Task<int> GetSampleSetCountAsync();
   Task<IEnumerable<int>> GetAllSampleSetIdsAsync();
   Task<int> GetIndexableSampleSetCountAsync();
   Task<int> GetSuspendedSampleSetCountAsync();
-
+  Task<SampleSet> GetSampleSetByIdAsync(int id);
 }
